requirements.txt
aiohttp==3.10.5
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
frozenlist==1.4.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
openai==0.28.0
packaging==24.1
pydantic==2.9.1
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.11.1